Run tests with karma-mocha
