+++ package.json
@@ -59,7 +59,7 @@
     "karma": "^6.3.9",
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^2.0.1",
-    "karma-spec-reporter": "^0.0.32",
+    "karma-mocha-reporter": "^2.2.5",
     "karma-webpack": "^5.0.0",
     "lint-staged": "^11.2.6",
     "mocha": "^9.1.3",
