+++ test/setup.ts
@@ -1,3 +1,8 @@
+import { use as chaiUse } from 'chai';
+import * as chaiAsPromised from 'chai-as-promised';
+
 import { init } from '../lib';
 
+chaiUse(chaiAsPromised);
 before(init);
+console.log('yoloooo')
