+++ tsconfig.json
@@ -5,12 +5,8 @@
       "strict": true,
       "sourceMap": true,
       "allowJs": true,
-      "outDir": "build/",
   },
   "include": [
-    "./test/**/*",
-    // these are not directly referenced by the tests, but are still needed
-    "./lib/openpgp.js",
-    "./lib/index.js"
+    "./test/**/*"
   ]
 }
